Refactor docker-specific analysis to new module
Currently Tern only supports docker-related analysis. In the future,
however, the goal is to support different types of container formats.
In order to make this integration easier, this commit refactors docker
analysis and reporting into its own module. A few notable changes
introduced by this commit:

1) Created tern/analyze/docker directory for docker-related analysis
   files.

2) Moved docker-specific functions from tern/report/report.py to
   tern/analyze/docker/run.py.

3) Moved tern/report/analyze.py, tern/utils/container.py, and
   tern/utils/dockerfile.py under tern/analyze/docker.

4) Updated the documentation related to refactoring mentioned
   in the bullets above.

5) Updated all import statements for any references to the
   refactored functions and files.

Works towards #262
