+++ tern/analyze/docker/dockerfile.py
