+++ tern/report/report.py
@@ -7,7 +7,6 @@
 Create a report
 """
 
-import docker
 import logging
 import os
 import shutil
@@ -19,18 +18,13 @@ from stevedore.exception import NoMatches
 
 from tern.report import errors
 from tern.report import formats
-from tern.report.analyze import analyze_docker_image
-from tern.utils import container
+from tern.analyze.docker import container
 from tern.utils import constants
 from tern.utils import cache
 from tern.utils import general
 from tern.utils import rootfs
 from tern.classes.docker_image import DockerImage
-from tern.classes.image import Image
-from tern.classes.image_layer import ImageLayer
 from tern.classes.notice import Notice
-from tern.classes.package import Package
-from tern.helpers import common
 import tern.helpers.docker as dhelper
 
 # global logger
@@ -140,37 +134,6 @@ def load_full_image(image_tag_string):
     return test_image
 
 
-def get_dockerfile_packages():
-    '''Given a Dockerfile return an approximate image object. This is mosty
-    guess work and shouldn't be relied on for accurate information. Add
-    Notice messages indicating as such:
-        1. Create an image with a placeholder repotag
-        2. For each RUN command, create a package list
-        3. Create layer objects with incremental integers and add the package
-        list to that layer with a Notice about parsing
-        4. Return stub image'''
-    stub_image = Image('easteregg:cookie')
-    layer_count = 0
-    for inst in dhelper.docker_commands:
-        if inst[0] == 'RUN':
-            layer_count = layer_count + 1
-            layer = ImageLayer(layer_count)
-            install_commands, msg = common.filter_install_commands(inst[1])
-            if msg:
-                layer.origins.add_notice_to_origins(
-                    inst[1], Notice(msg, 'info'))
-            pkg_names = []
-            for command in install_commands:
-                pkg_names.append(common.get_installed_package_names(command))
-            for pkg_name in pkg_names:
-                pkg = Package(pkg_name)
-                # shell parser does not parse version pins yet
-                # when that is enabled, Notices for no versions need to be
-                # added here
-                layer.add_package(pkg)
-    return stub_image
-
-
 def generate_report(args, *images):
     '''Generate a report based on the command line options'''
     if args.report_format:
@@ -201,107 +164,3 @@ def report_out(args, *images):
         write_report(report, args)
     else:
         print(report)
-
-
-def check_docker_daemon():
-    '''Check if the Docker daemon is running. If not, exit gracefully'''
-    try:
-        docker.from_env()
-    except IOError as error:
-        logger.error('Docker daemon is not running: %s',
-                     error.output.decode('utf-8'))
-        sys.exit()
-
-
-def execute_dockerfile(args):
-    '''Execution path if given a dockerfile'''
-    check_docker_daemon()
-    logger.debug('Setting up...')
-    setup(dockerfile=args.dockerfile)
-    # attempt to build the image
-    logger.debug('Building Docker image...')
-    # placeholder to check if we can analyze the full image
-    completed = True
-    build, _ = dhelper.is_build()
-    if build:
-        # attempt to get built image metadata
-        image_tag_string = dhelper.get_dockerfile_image_tag()
-        full_image = load_full_image(image_tag_string)
-        if full_image.origins.is_empty():
-            # image loading was successful
-            # Add an image origin here
-            full_image.origins.add_notice_origin(
-                formats.dockerfile_image.format(dockerfile=args.dockerfile))
-            # analyze image
-            analyze_docker_image(full_image, args.redo, True)
-        else:
-            # we cannot load the full image
-            logger.warning('Cannot retrieve full image metadata')
-            completed = False
-        # clean up image
-        container.remove_image(full_image.repotag)
-        if not args.keep_wd:
-            clean_image_tars(full_image)
-    else:
-        # cannot build the image
-        logger.warning('Cannot build image')
-        completed = False
-    # check if we have analyzed the full image or not
-    if not completed:
-        # get the base image
-        logger.debug('Loading base image...')
-        base_image = load_base_image()
-        if base_image.origins.is_empty():
-            # image loading was successful
-            # add a notice stating failure to build image
-            base_image.origins.add_notice_to_origins(
-                args.dockerfile, Notice(
-                    formats.image_build_failure, 'warning'))
-            # analyze image
-            analyze_docker_image(base_image, args.redo)
-        else:
-            # we cannot load the base image
-            logger.warning('Cannot retrieve base image metadata')
-        # run through commands in the Dockerfile
-        logger.debug('Parsing Dockerfile to generate report...')
-        stub_image = get_dockerfile_packages()
-        # clean up image
-        container.remove_image(base_image.repotag)
-        if not args.keep_wd:
-            clean_image_tars(base_image)
-    # generate report based on what images were created
-    if completed:
-        report_out(args, full_image)
-    else:
-        report_out(args, base_image, stub_image)
-    logger.debug('Teardown...')
-    teardown()
-    if not args.keep_wd:
-        clean_working_dir(args.bind_mount)
-
-
-def execute_docker_image(args):
-    '''Execution path if given a Docker image'''
-    check_docker_daemon()
-    logger.debug('Setting up...')
-    setup(image_tag_string=args.docker_image)
-    # attempt to get built image metadata
-    full_image = load_full_image(args.docker_image)
-    if full_image.origins.is_empty():
-        # image loading was successful
-        # Add an image origin here
-        full_image.origins.add_notice_origin(
-            formats.docker_image.format(imagetag=args.docker_image))
-        # analyze image
-        analyze_docker_image(full_image, args.redo)
-        # generate report
-        report_out(args, full_image)
-    else:
-        # we cannot load the full image
-        logger.warning('Cannot retrieve full image metadata')
-    if not args.keep_wd:
-        clean_image_tars(full_image)
-    logger.debug('Teardown...')
-    teardown()
-    if not args.keep_wd:
-        clean_working_dir(args.bind_mount)
